adds species to the schema

+++ src/main/java/uk/ac/ebi/pride/cluster/search/service/ClusterIndexService.java
@@ -9,13 +9,15 @@ import uk.ac.ebi.pride.cluster.search.service.repository.SolrClusterRepository;
 
 import java.util.Collection;
 import java.util.Collections;
+import java.util.List;
 
 /**
  * Created by jdianes on 04/11/2014.
+ *
  * @author ntoro
  */
 @Service
-public class ClusterIndexService {
+public class ClusterIndexService implements IClusterIndexService {
 
     private static Logger logger = LoggerFactory.getLogger(ClusterIndexService.class);
 
@@ -29,12 +31,14 @@ public class ClusterIndexService {
         this.solrClusterRepository = solrClusterRepository;
     }
 
+    @Override
     @Transactional
-    public void save(Cluster cluster){
+    public void save(Cluster cluster) {
         Collection<Cluster> clusterCollection = Collections.singletonList(cluster);
         save(clusterCollection);
     }
 
+    @Override
     @Transactional
     public void save(Iterable<Cluster> clusters) {
         if (clusters != null && clusters.iterator().hasNext()) {
@@ -44,24 +48,44 @@ public class ClusterIndexService {
         }
     }
 
+    @Override
     @Transactional
-    public void deleteAll() {
-        solrClusterRepository.deleteAll();
+    public void delete(long clusterId) {
+        solrClusterRepository.delete(clusterId);
     }
 
+    @Override
     @Transactional
-    public void delete(Cluster cluster){
-        solrClusterRepository.delete(cluster);
+    public void delete(Cluster cluster) {
+        if(cluster != null){
+            solrClusterRepository.delete(cluster);
+        }  else {
+            logger.info("No cluster to delete");
+        }
     }
 
+    @Override
+    @Transactional
+    public void delete(List<Long> clusterIds) {
+        Iterable<Cluster> clusters = solrClusterRepository.findAll(clusterIds);
+        delete(clusters);
+    }
+
+    @Override
     @Transactional
     public void delete(Iterable<Cluster> clusters) {
         if (clusters != null && clusters.iterator().hasNext()) {
             solrClusterRepository.delete(clusters);
         } else {
             logger.info("No clusters to delete");
-
         }
     }
+
+    @Override
+    @Transactional
+    public void deleteAll() {
+        solrClusterRepository.deleteAll();
+    }
+
 
 }
